+++ src/main/java/eu/deltacraft/deltacraftteams/types/IsTeamOwnerResponse.kt
@@ -1,0 +1,6 @@
+package eu.deltacraft.deltacraftteams.types
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class IsTeamOwnerResponse(val content: Boolean)
